Simplify and improve assembly resolution
Instead of generating references to be passed to the compiler,
generate references to be passed to MSBuild's reference
resolution, since it'll take care of deduplicating and
resolving conflicts.

Add the bin directory to the resolve path so that core assemblies
that are not explicitly part of addins can still be resolved.

Reference a number of common core assemblies by default.